Add gauges with strobe reading.
This change adds a `GaugeReporter`, which may be used to automatically
strobe gauges and report values back to the metric implementation. An
application will generally only require one instance, but this is not a
restriction.  The `GaugeReporter` is thread safe and can be modified at
any point of time and shared across threads.

Gauges may be added to a reporter using the `addGauge` method, which
takes a name and reading provider. The reading provider must be thread
safe and should not block for long periods of time, as this will hold up
the reporter thread.

This change also adds in a number of very useful, common-scenario gauges
covering the most common JVM metrics, such as threads, memory, and
garbage collection statistics. These sets may be added with the
`addGauges` method, such as `addGauges(jvmGauges())`.